Finished refactoring of captureOrMatch to engine_util; added test of match/capture in socket.io test
