+++ lib/engine_http.js
@@ -7,10 +7,8 @@
 const async = require('async');
 const _ = require('lodash');
 const request = require('request');
-const jsonpath = require('JSONPath');
 const debug = require('debug')('http');
 const debugResponse = require('debug')('http:response');
-const debugCapture = require('debug')('http:capture');
 const VERSION = require('../package.json').version;
 const USER_AGENT = 'artillery ' + VERSION + ' (https://artillery.io)';
 const engineUtil = require('./engine_util');
@@ -19,14 +17,6 @@ const http = require('http');
 const https = require('https');
 const fs = require('fs');
 const filtrex = require('filtrex');
-const cheerio = require('cheerio');
-
-let xmlCapture;
-try {
-  xmlCapture = require('artillery-xml-capture');
-} catch (e) {
-  xmlCapture = null;
-}
 
 module.exports = HttpEngine;
 
@@ -255,7 +245,7 @@ HttpEngine.prototype.step = function step(requestSpec, ee, opts) {
                 return callback(err, context);
               }
               if (params.capture || params.match) {
-                captureOrMatch(params, res, context, function(err, result) {
+                engineUtil.captureOrMatch(params, res, context, function(err, result) {
                   if (err) {
                     return callback(null, context);
                   }
@@ -382,235 +372,4 @@ function lowcaseKeys(h) {
   return _.transform(h, function(result, v, k) {
     result[k.toLowerCase()] = v;
   });
-}
-
-/*
- * Given a response object determine if it's JSON
- */
-function isJSON(res) {
-  debug('isJSON: content-type = %s', res.headers['content-type']);
-  return (res.headers['content-type'] &&
-          /^application\/json/.test(res.headers['content-type']));
-}
-
-/*
- * Given a response object determine if it's some kind of XML
- */
-function isXML(res) {
-  return (res.headers['content-type'] &&
-          (/^[a-zA-Z]+\/xml/.test(res.headers['content-type']) ||
-           /^[a-zA-Z]+\/[a-zA-Z]+\+xml/.test(res.headers['content-type'])));
-
-}
-
-/*
- * Wrap JSON.parse in a callback
- */
-function parseJSON(body, callback) {
-  let r = null;
-  let err = null;
-
-  try {
-    if (typeof body === 'string') {
-      r = JSON.parse(body);
-    } else {
-      r = body;
-    }
-  } catch(e) {
-    err = e;
-  }
-
-  return callback(err, r);
-}
-
-function dummyParser(body, callback) {
-  return callback(null, body);
-}
-
-// doc is a JSON object
-function extractJSONPath(doc, expr) {
-  let results = jsonpath.eval(doc, expr);
-  if (results.length > 1) {
-    return results[randomInt(0, results.length - 1)];
-  } else {
-    return results[0];
-  }
-}
-
-// doc is a string or an object (body parsed by Request when headers indicate JSON)
-function extractRegExp(doc, expr, opts) {
-  let group = opts.group;
-  let str;
-  if (typeof doc === 'string') {
-    str = doc;
-  } else {
-    str = JSON.stringify(doc); // FIXME: not the same string as the one we got from the server
-  }
-  let rx = new RegExp(expr);
-  let match = rx.exec(str);
-  if(group && match[group]) {
-    return match[group];
-  } else if (match[0]) {
-    return match[0];
-  } else {
-    return '';
-  }
-}
-
-function extractCheerio(doc, expr, opts) {
-  let $ = cheerio.load(doc);
-  let els = $(expr);
-  let i = 0;
-  if (typeof opts.index !== 'undefined') {
-    if (opts.index === 'random') {
-      i = Math.ceil(Math.random() * els.get().length - 1);
-    } else if (opts.index === 'last') {
-      i = els.get().length() - 1;
-    } else if (typeof Number(opts.index) === 'number') {
-      i = Number(opts.index);
-    }
-  }
-  return els.slice(i, i + 1).attr(opts.attr);
-}
-
-function extractHeader(headers, headerName) {
-  return headers[headerName];
-}
-
-function dummyExtractor() {
-  return '';
-}
-
-function randomInt (low, high) {
-  return Math.floor(Math.random() * (high - low + 1) + low);
-}
-
-HttpEngine.captureOrMatch = captureOrMatch;
-
-//
-// Calls done() with:
-// {captures: { var: value }, matches: { var: {expected: '', got: ''} }}
-//
-function captureOrMatch(params, response, context, done) {
-  let specs = _.concat(
-    _.get(params, 'capture', []),
-    _.get(params, 'match', []));
-
-  let result = {
-    captures: {},
-    matches: {}
-  };
-
-  async.eachSeries(
-    specs,
-    function(spec, next) {
-      let parsedSpec = parseSpec(spec, response);
-      let parser = parsedSpec.parser;
-      let extractor = parsedSpec.extractor;
-      let expr = parsedSpec.expr;
-
-      // are we looking at body or headers:
-      var content = response.body;
-      if (spec.header) {
-        content = response.headers;
-      }
-
-      parser(content, function(err, doc) {
-        if (err) {
-          return next(err, null);
-        }
-
-        let extractedValue = extractor(doc, expr, spec);
-
-        if (spec.value) {
-          // this is a match spec
-          let expected = template(spec.value, context);
-          debug('match: %s, expected: %s, got: %s', expr, expected, extractedValue);
-          if (extractedValue !== expected) {
-            result.matches[expr] = {
-              success: false,
-              expected: expected,
-              got: extractedValue,
-              expression: expr,
-              strict: spec.strict
-            };
-          } else {
-            result.matches.expr = {
-              success: true,
-              expected: expected,
-              expression: expr
-            };
-          }
-          return next(null);
-        }
-
-        if (spec.as) {
-          // this is a capture
-          debugCapture('capture: %s = %s', spec.as, extractedValue);
-          result.captures[spec.as] = extractedValue;
-          if (spec.transform) {
-            let transformedValue = engineUtil.evil(
-              result.captures,
-              spec.transform);
-
-            debugCapture('transform: %s = %s', spec.as, result.captures[spec.as]);
-            result.captures[spec.as] = transformedValue !== null ? transformedValue : extractedValue;
-          }
-        }
-
-        return next(null);
-      });
-    },
-    function(err) {
-      if (err) {
-        return done(err, null);
-      } else {
-        return done(null, result);
-      }
-    });
-}
-
-function parseSpec(spec, response) {
-  let parser;
-  let extractor;
-  let expr;
-
-  if (spec.json) {
-    parser = parseJSON;
-    extractor = extractJSONPath;
-    expr = spec.json;
-  } else if (xmlCapture && spec.xpath) {
-    parser = xmlCapture.parseXML;
-    extractor = xmlCapture.extractXPath;
-    expr = spec.xpath;
-  } else if (spec.regexp) {
-    parser = dummyParser;
-    extractor = extractRegExp;
-    expr = spec.regexp;
-  } else if (spec.header) {
-    parser = dummyParser;
-    extractor = extractHeader;
-    expr = spec.header;
-  } else if (spec.selector) {
-    parser = dummyParser;
-    extractor = extractCheerio;
-    expr = spec.selector;
-  } else {
-    if (isJSON(response)) {
-      parser = parseJSON;
-      extractor = extractJSONPath;
-      expr = spec.json;
-    } else if (xmlCapture && isXML(response)) {
-      parser = xmlCapture.parseXML;
-      extractor = xmlCapture.extractXPath;
-      expr = spec.xpath;
-    } else {
-      // We really don't know what to do here.
-      parser = dummyParser;
-      extractor = dummyExtractor;
-      expr = '';
-    }
-  }
-
-  return { parser: parser, extractor: extractor, expr: expr };
 }
