+++ lib/engine_socketio.js
@@ -59,7 +59,7 @@ function processResponse(ee, data, response, context, callback) {
   let fauxResponse = {body: JSON.stringify(data)};
 
   // Handle the capture or match clauses...
-  EngineHttp.captureOrMatch(response, fauxResponse, context, function(err, result) {
+  engineUtil.captureOrMatch(response, fauxResponse, context, function(err, result) {
     // Were we unable to invoke captureOrMatch?
     if (err) {
       debug(data);
@@ -98,6 +98,7 @@ function processResponse(ee, data, response, context, callback) {
 
       // Increment the success count...
       context._successCount++;
+
       return callback(null, context);
     }
   });
