+++ test/scripts/hello_socketio.json
@@ -21,7 +21,9 @@
         {"emit": { "channel": "echo", "data": "{\"key\":\"value\"}", "response": { "channel": "echoed", "data": "{\"key\":\"value\"}"} }},
         {"think": 1},
         {"emit": { "channel": "echo", "data": "random number: {{ $randomNumber(1, 20) }}" }},
-        {"emit": { "channel": "echo", "data": "random string: {{ $randomString(10) }}" }},
+        {"emit": { "channel": "echo", "data": {"key": "{{ $randomString(10) }}"}, "response": { "channel": "echoed", "capture": {"json": "$.key", "as": "rand" } }}},
+        {"think": 1},
+        {"emit": { "channel": "echo", "data": {"key": "{{ rand }}"}, "response": { "channel": "echoed", "match": {"json": "$.key", "value": "{{ rand }}" } }}},
         {"think": 1}
       ]
     }
